Refactor formulating index.html file logic so we can use Ngrok to test it
